# docs/API_SPEC.yaml
openapi: 3.1.0
info:
  title: SKAI-ToDo API
  version: 2.0.0
  description: REST and WebSocket API for SKAI-ToDo v2.0
servers:
  - url: https://api.skai-todo.local/v1
    description: Local Development Server
paths:
  /auth/register:
    post:
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
                name:
                  type: string
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'

  /auth/login:
    post:
      summary: Authenticate user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Successful authentication
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'

  /tasks:
    get:
      summary: Get all tasks for the user/organization
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
    post:
      summary: Create a new task manually
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskInput'
      responses:
        '201':
          description: Task created

  /tasks/voice:
    post:
      summary: Process voice dictation to task via n8n
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                audio:
                  type: string
                  format: binary
      responses:
        '202':
          description: Processing started (client will be notified via WebSocket)

  /webhooks/stripe:
    post:
      summary: Stripe billing event webhook
      responses:
        '200':
          description: OK

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        user:
          $ref: '#/components/schemas/User'
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: string
        role:
          type: string
    Task:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        quadrant:
          type: string
          enum: [Q1, Q2, Q3, Q4]
        status:
          type: string
        dueDate:
          type: string
          format: date-time
    TaskInput:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        quadrant:
          type: string
          enum: [Q1, Q2, Q3, Q4]